Respecialized size_t template

+++ data/result_container.cpp
@@ -1,10 +1,12 @@
+#include <iostream>
+
 #include "result_container.hpp"
 
 namespace data
 {
 
 template<>
-void ResultContainer<int>::Merge(const ResultContainer<int>& other)
+void ResultContainer<size_t>::Merge(const ResultContainer<size_t>& other)
 {
 	mValue += other.Value();
 }
